// packages/auction-house-nft-hooks/src/graph-queries/zora-graph-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: any;
  BigInt: any;
  Bytes: any;
};




export type Ask = {
  __typename?: 'Ask';
  /** <tokenId>-<ownerAddress> */
  id: Scalars['ID'];
  /** The Media associated with the Ask */
  media: Media;
  /** Transaction hash the ask was created at */
  transactionHash: Scalars['String'];
  /** The Currency of the Ask */
  currency: Currency;
  /** The amount of Currency of the Ask */
  amount: Scalars['BigInt'];
  /** The owner of the Ask */
  owner: User;
  /** The timestamp of the block the Ask was created in */
  createdAtTimestamp: Scalars['BigInt'];
  /** The number of the block the Ask created in */
  createdAtBlockNumber: Scalars['BigInt'];
};
// ...
export type Bid = {
  __typename?: 'Bid';
  /** <token-id>-<bidderAddress> */
  id: Scalars['ID'];
  /** Transaction hash the bid was created at */
  transactionHash: Scalars['String'];
  /** The Media associated with the Bid */
  media: Media;
  /** The Currency of the Bid */
  currency: Currency;
  /** The amount of Currency of the Bid */
  amount: Scalars['BigInt'];
  /** The sellOnShare of the Bid */
  sellOnShare: Scalars['BigInt'];
  /** The bidder of the Bid */
  bidder: User;
  /** The recipient of Media if the Bid is accepted */
  recipient: User;
  /** The timestamp of the block the Bid was created in */
  createdAtTimestamp: Scalars['BigInt'];
  /** The number of the block the Bid was created in */
  createdAtBlockNumber: Scalars['BigInt'];
};
// ...
export type Currency = {
  __typename?: 'Currency';
  /** The address of the Currency */
  id: Scalars['ID'];
  /** The name of the Currency */
  name: Scalars['String'];
  /** The symbol of the Currency */
  symbol: Scalars['String'];
  /** The decimals of the Currency */
  decimals?: Maybe<Scalars['Int']>;
  /** Total Bid Liquidity of the Currency on all Zora Media */
  liquidity: Scalars['BigInt'];
  /** The active Bids denominated in the Currency */
  activeBids?: Maybe<Array<Bid>>;
  /** The active Asks denominated in the Currency */
  activeAsks?: Maybe<Array<Ask>>;
  /** The InactiveBids denominated in the Currency */
  inactiveBids?: Maybe<Array<InactiveBid>>;
  /** The InactiveAsks denominated in the Currency */
  inactiveAsks?: Maybe<Array<InactiveAsk>>;
};
// ...
export type InactiveReserveAuctionBid = {
  __typename?: 'InactiveReserveAuctionBid';
  /** <auctionId>-<txHash>-<logIndex> */
  id: Scalars['ID'];
  /** The transaction hash the reserve auction big was created at */
  transactionHash: Scalars['String'];
  /** The Reserve auction associated with the Bid */
  reserveAuction: ReserveAuction;
  /** The Currency of the Bid */
  currency: Currency;
  /** The amount of the Bid */
  amount: Scalars['BigInt'];
  /** The bidder of the Bid */
  bidder: User;
  /** The type of bid (active, refunded, final) */
  bidType: ReserveAuctionBidType;
  /** The timestamp of the block the bid was inactivated at (via outbid, cancellation, winning bid) */
  bidInactivatedAtTimestamp: Scalars['BigInt'];
  /** The number of the block the bid was inactivated at (via outbid, cancellation, winning bid) */
  bidInactivatedAtBlockNumber: Scalars['BigInt'];
  /** The bid is buyNow */
  buyNowRequested: Scalars['Boolean'];
  /** The bid is buyNow */
  buyNowActual: Scalars['Boolean'];
  /** The timestamp of the block the Bid was created in */
  createdAtTimestamp: Scalars['BigInt'];
  /** The number of the block the Bid was created in */
  createdAtBlockNumber: Scalars['BigInt'];
};
// ...
export type Media = {
  __typename?: 'Media';
  /** The tokenId on the Zora Media Contract */
  id: Scalars['ID'];
  /** The transaction hash the media was created at */
  transactionHash: Scalars['String'];
  /** The current owner of the Media */
  owner: User;
  /** The creator of the Media */
  creator: User;
  /** The previous owner of the Zora Media's Market */
  prevOwner: User;
  /** The approved user of the Media */
  approved?: Maybe<User>;
  /** The sha256 hash of the media's content */
  contentHash: Scalars['Bytes'];
  /** The sha256 hash of the media's metadata */
  metadataHash: Scalars['Bytes'];
  /** The uri of the content */
  contentURI: Scalars['String'];
  /** The uri of the metadata */
  metadataURI: Scalars['String'];
  /** The bid share for the current owner of the Media */
  ownerBidShare: Scalars['BigInt'];
  /** The bid share for the creator of the Media */
  creatorBidShare: Scalars['BigInt'];
  /** The bid share for the previous owner of the Media's market */
  prevOwnerBidShare: Scalars['BigInt'];
  /** The timestamp of the block the Media was minted in */
  createdAtTimestamp: Scalars['BigInt'];
  /** The number of the block the Media was minted in */
  createdAtBlockNumber: Scalars['BigInt'];
  /** The timestamp of the block the Media was burned in */
  burnedAtTimeStamp?: Maybe<Scalars['BigInt']>;
  /** The number of the block the Media was burned in */
  burnedAtBlockNumber?: Maybe<Scalars['BigInt']>;
  /** The current Ask of the Media */
  currentAsk?: Maybe<Ask>;
  /** The current Bids on the Media */
  currentBids?: Maybe<Array<Bid>>;
  /** The InactiveAsks of the Media */
  inactiveAsks?: Maybe<Array<InactiveAsk>>;
  /** The InactiveBids of the Media */
  inactiveBids?: Maybe<Array<InactiveBid>>;
  /** The transfers of the Media */
  transfers?: Maybe<Array<Transfer>>;
  /** The ReserveAuctions of the Media */
  reserveAuctions?: Maybe<Array<ReserveAuction>>;
};
// ...
export type ReserveAuction = {
  __typename?: 'ReserveAuction';
  /** ID of the auction from contract, autoincrementing int */
  id: Scalars['ID'];
  /** Transaction hash where the reserve auction was created */
  transactionHash: Scalars['String'];
  /** The originating token contract for this auction */
  tokenContract: Scalars['String'];
  /** The token ID for this auction */
  tokenId: Scalars['BigInt'];
  /** <tokenContract>-<tokenId> */
  token: Scalars['String'];
  /** The media for the auction, if it is a zora NFT */
  media?: Maybe<Media>;
  /** Whether or not the auction has been approved by the curator */
  approved: Scalars['Boolean'];
  /** The length of time the auction is intended to run for, after the first bid is made */
  duration: Scalars['BigInt'];
  /**
   * The expected end of auction timestamp, which can change if bids were placed
   * within the last 15 minutes, and is not set until the first bid is placed
   */
  expectedEndTimestamp?: Maybe<Scalars['BigInt']>;
  /** The time the first bid was placed */
  firstBidTime?: Maybe<Scalars['BigInt']>;
  /** The minimum price of the first bid */
  reservePrice: Scalars['BigInt'];
  /** The minimum price currency of the first bid */
  reserveAndBuyNowCurrency: Currency;
  /** The sale percentage to send to the curator */
  curatorFeePercentage: Scalars['Int'];
  /** The address that should receive the funds once the NFT is sold */
  tokenOwner: User;
  /** The address of the auction's curator */
  curator: User;
  /** The address of the ERC-20 currency to run the auction with, or 0x0 for ETH */
  auctionCurrencies?: Maybe<Array<Currency>>;
  status: ReserveAuctionStatus;
  /** The current bid on this auction */
  currentBid?: Maybe<ReserveAuctionBid>;
  /** The buyNow price in reserve currency */
  buyNowPrice: Scalars['BigInt'];
  /** The previous bids on this auction */
  previousBids?: Maybe<Array<InactiveReserveAuctionBid>>;
  /** The time the auction was approved */
  approvedTimestamp?: Maybe<Scalars['BigInt']>;
  /** The number of the block the auction was Approved */
  approvedBlockNumber?: Maybe<Scalars['BigInt']>;
  /** The timestamp of the block the Transfer was created in */
  createdAtTimestamp: Scalars['BigInt'];
  /** The number of the block the Transfer was created in */
  createdAtBlockNumber: Scalars['BigInt'];
  /** The timestamp at which the auction end function was called */
  finalizedAtTimestamp?: Maybe<Scalars['BigInt']>;
  /** The block number at which the auction end function was called */
  finalizedAtBlockNumber?: Maybe<Scalars['BigInt']>;
};
// ...
export type ReserveAuctionBid = {
  __typename?: 'ReserveAuctionBid';
  /** <auctionId>-<txHash>-<logIndex> */
  id: Scalars['ID'];
  /** The transaction hash the reserve auction big was created at */
  transactionHash: Scalars['String'];
  /** The Reserve auction associated with the Bid */
  reserveAuction: ReserveAuction;
  /** The Currency of the Bid */
  currency: Currency;
  /** The amount of the Bid */
  amount: Scalars['BigInt'];
  /** The bidder of the Bid */
  bidder: User;
  /** The type of bid (active, refunded, final) */
  bidType: ReserveAuctionBidType;
  /** The bid is buyNow */
  buyNowRequested: Scalars['Boolean'];
  /** The bid is buyNow */
  buyNowActual: Scalars['Boolean'];
  /** The timestamp of the block the Bid was created in */
  createdAtTimestamp: Scalars['BigInt'];
  /** The number of the block the Bid was created in */
  createdAtBlockNumber: Scalars['BigInt'];
};

export enum ReserveAuctionBidType {
  Active = 'Active',
  Refunded = 'Refunded',
  Final = 'Final'
}
// ...
export enum ReserveAuctionStatus {
  Pending = 'Pending',
  Active = 'Active',
  Canceled = 'Canceled',
  Finished = 'Finished'
}
// ...
export type User = {
  __typename?: 'User';
  /** Ethereum Address */
  id: Scalars['ID'];
  /** Users that have been granted `ApprovalForAll` Media of the User's Collection */
  authorizedUsers?: Maybe<Array<User>>;
  /** The Media the User owns */
  collection: Array<Media>;
  /** The Media the User created */
  creations: Array<Media>;
  /** The active Bids made by the User */
  currentBids?: Maybe<Array<Bid>>;
};
// ...
export type CurrencyShortFragment = (
  { __typename?: 'Currency' }
  & Pick<Currency, 'id' | 'name' | 'symbol' | 'decimals'>
);

export type PreviousReserveBidFragment = (
  { __typename?: 'InactiveReserveAuctionBid' }
  & Pick<InactiveReserveAuctionBid, 'id' | 'transactionHash' | 'createdAtTimestamp' | 'amount' | 'bidType' | 'buyNowRequested' | 'buyNowActual' | 'bidInactivatedAtTimestamp' | 'bidInactivatedAtBlockNumber'>
  & { bidder: (
    { __typename?: 'User' }
    & Pick<User, 'id'>
  ), currency: (
    { __typename?: 'Currency' }
    & CurrencyShortFragment
  ) }
);

export type CurrentReserveBidFragment = (
  { __typename?: 'ReserveAuctionBid' }
  & Pick<ReserveAuctionBid, 'id' | 'transactionHash' | 'createdAtTimestamp' | 'amount' | 'bidType' | 'buyNowRequested' | 'buyNowActual'>
  & { bidder: (
    { __typename?: 'User' }
    & Pick<User, 'id'>
  ), currency: (
    { __typename?: 'Currency' }
    & CurrencyShortFragment
  ) }
);

export type ReserveAuctionPartialFragment = (
  { __typename?: 'ReserveAuction' }
  & Pick<ReserveAuction, 'id' | 'tokenId' | 'tokenContract' | 'transactionHash' | 'status' | 'approved' | 'reservePrice' | 'buyNowPrice' | 'firstBidTime' | 'token' | 'createdAtTimestamp' | 'approvedTimestamp' | 'curatorFeePercentage' | 'duration' | 'expectedEndTimestamp' | 'finalizedAtTimestamp'>
  & { reserveAndBuyNowCurrency: (
    { __typename?: 'Currency' }
    & CurrencyShortFragment
  ), curator: (
    { __typename?: 'User' }
    & Pick<User, 'id'>
  ), tokenOwner: (
    { __typename?: 'User' }
    & Pick<User, 'id'>
  ), auctionCurrencies?: Maybe<Array<(
    { __typename?: 'Currency' }
    & CurrencyShortFragment
  )>>, currentBid?: Maybe<(
    { __typename?: 'ReserveAuctionBid' }
    & CurrentReserveBidFragment
  )>, previousBids?: Maybe<Array<(
    { __typename?: 'InactiveReserveAuctionBid' }
    & PreviousReserveBidFragment
  )>> }
);